+++ buildSrc/src/main/kotlin/Configurations.kt
@@ -36,24 +36,24 @@ object Versions {
     }
 
     object Sdk {
-        const val classpath = "3.6.1"
-        const val buildTools = "29.0.3"
-        const val compile = 29
+        const val classpath = "4.0.1"
+        const val buildTools = "30.0.2"
+        const val compile = 30
         const val min = 21
-        const val target = 29
+        const val target = 30
     }
 
     val java = JavaVersion.VERSION_1_8
 
     object Kotlin {
-        const val core = "1.3.70"
+        const val core = "1.4.0"
     }
 
     object AndroidX {
         const val annotation = "1.1.0"
     }
 
-    const val mockito = "3.3.0"
+    const val mockito = "3.5.9"
 }
 
 object Dependencies {
@@ -67,7 +67,6 @@ object Dependencies {
 
     object Kotlin {
         const val classpath = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.Kotlin.core}"
-        const val stdlib = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.Kotlin.core}"
         const val test = "org.jetbrains.kotlin:kotlin-test-junit:${Versions.Kotlin.core}"
     }