+++ CHANGELOG.md
@@ -2,6 +2,15 @@ CHANGELOG
 ---------
 
 #### Next Release
+- Updated dependencies:
+  - Gradle to 6.6.1
+  - Android Gradle plugin to 4.0.1
+  - Android build tool to 30.0.2
+  - Compile and target SDK to 30
+  - Kotlin to 1.4.0
+  - Mockito to 3.5.9
+
+#### 0.2.0
 - Used Executor instead of coroutine for FileLogger
 - Updated dependencies:
   - Gradle to 6.2.2